// src/api/register/RegistorValidation.ts
export class RegisterValidation {
    InvalidName: boolean = false;
    InvalidNickname: boolean = false;
    InvalidEmail: boolean = false;
    InvalidPassword: boolean = false;
    InvalidConfirmation: boolean = false;
    InvalidPhone: boolean = false;
    InvalidType: boolean = false;
    error: string = '';
    showPassword: boolean = false;
    showConfirmPass: boolean = false;
  }
  